pyproject.toml
[project]
name = "pydew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bombercorny", email = "claudius.hini@hotmail.de" }
]
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.5",
    "pytmx>=3.32",
    "ruff>=0.12.2",
]

[project.scripts]
pydew = "pydew:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
